Ponderada 3/backend/Flask app (docker)/requirements.txt
flask
Flask-SQLAlchemy
Flask-JWT-Extended
requests
Flask-WTF
psycopg2-binary
asyncpg
Pillow
rembg